chore #58: 방 개설 코드 폴더 이동
