+++ ody_flutter/lib/screens/gathering_creator/gathering_creator_screen.dart
@@ -5,11 +5,11 @@ import "package:ody_flutter/components/ody_top_bar.dart";
 import "package:ody_flutter/data/network/base/base_service.dart";
 import "package:ody_flutter/data/network/service/gathering_service_impl.dart";
 import "package:ody_flutter/data/repository/gathering_repository_impl.dart";
-import "package:ody_flutter/screens/gatherings/gathering_creator_view_model.dart";
-import "package:ody_flutter/screens/gatherings/screens/gathering_date_screen.dart";
-import "package:ody_flutter/screens/gatherings/screens/gathering_location_screen.dart";
-import "package:ody_flutter/screens/gatherings/screens/gathering_time_screen.dart";
-import "package:ody_flutter/screens/gatherings/screens/gathering_title_screen.dart";
+import "package:ody_flutter/screens/gathering_creator/gathering_creator_view_model.dart";
+import "package:ody_flutter/screens/gathering_creator/screens/gathering_date_screen.dart";
+import "package:ody_flutter/screens/gathering_creator/screens/gathering_location_screen.dart";
+import "package:ody_flutter/screens/gathering_creator/screens/gathering_time_screen.dart";
+import "package:ody_flutter/screens/gathering_creator/screens/gathering_title_screen.dart";
 import "package:smooth_page_indicator/smooth_page_indicator.dart";
 
 class GatheringCreatorScreen extends StatelessWidget {
