+++ ody_flutter/lib/screens/gathering_creator/gathering_creator_view_model.dart
