+++ ody_flutter/lib/screens/gathering_creator/model/location.dart
