+++ ody_flutter/lib/screens/gathering_creator/screens/gathering_location_screen.dart
@@ -4,10 +4,12 @@ import "package:ody_flutter/assets/fonts/pretendard_fonts.dart";
 import "package:ody_flutter/components/ody_button.dart";
 import "package:ody_flutter/components/ody_highlight_text.dart";
 import "package:ody_flutter/components/ody_text_field.dart";
-import "package:ody_flutter/screens/gatherings/gathering_creator_view_model.dart";
+import "package:ody_flutter/config/routes.dart";
+import "package:ody_flutter/screens/gathering_creator/gathering_creator_view_model.dart";
+import "package:ody_flutter/screens/gathering_creator/model/location.dart";
 
-class GatheringTitleScreen extends StatelessWidget {
-  const GatheringTitleScreen({required this.viewModel, super.key});
+class GatheringLocationScreen extends StatelessWidget {
+  const GatheringLocationScreen({required this.viewModel, super.key});
 
   final GatheringCreatorViewModel viewModel;
 
@@ -22,8 +24,8 @@ class GatheringTitleScreen extends StatelessWidget {
               child: Column(
                 children: [
                   OdyHighlightText(
-                    text: "무슨 약속인가요?",
-                    highlightText: "약속",
+                    text: "오디서 만나시나요?",
+                    highlightText: "오디",
                     textStyle: PretendardFonts.bold24.copyWith(
                       color: CommonColors.gray_800,
                     ),
@@ -34,10 +36,27 @@ class GatheringTitleScreen extends StatelessWidget {
                   const SizedBox(
                     height: 32,
                   ),
-                  OdyTextField(
-                    textFieldType: OdyTextFieldType.textCounter,
-                    placeHolder: "약속을 입력해주세요",
-                    text: viewModel.text,
+                  GestureDetector(
+                    child: AbsorbPointer(
+                      child: OdyTextField(
+                        textFieldType: OdyTextFieldType.clearButton,
+                        placeHolder: "주소를 찾아보세요",
+                        text: viewModel.locationText,
+                      ),
+                    ),
+                    onTap: () async {
+                      final result = await Navigator.pushNamed(
+                        context,
+                        Routes.gatheringLocationSearch,
+                        arguments: Location(),
+                      );
+
+                      if (result != null) {
+                        viewModel.locationText.value =
+                            (result as Location).address ?? "";
+                        viewModel.isConfirmEnabled.value = true;
+                      }
+                    },
                   ),
                   const Spacer(),
                   OdyButton(
